third_party/blink/web_tests/fullscreen/full-screen-request-rejected-expected.txt

CONSOLE WARNING: Failed to execute 'requestFullscreen' on 'Element': API can only be initiated by a user gesture.
EVENT(webkitfullscreenerror)
END OF TEST

This tests that a full-screen request which is not user-initiated will be rejected.
